8. Sarcasm is lost in print

+++ main.h
@@ -5,7 +5,7 @@
 int _printf(const char *format, ...);
 int _print_string(int len, char *p);
 int _print_positif_int(int len, long int n);
-int _print_negatif_int(int len, int n);
+int _print_negatif_int(int len, long int n);
 int _print_NULL(int len);
 int _print_char(int len, const char *p);
 int _convert_ui_b(int len, long int n);
